editoast: move TrainId to models
